+++ app/actions/v2/services/service_binding_create.rb
@@ -2,6 +2,7 @@ require 'actions/services/database_error_service_resource_cleanup'
 require 'actions/services/locks/lock_check'
 require 'repositories/service_binding_event_repository'
 require 'jobs/v2/services/service_binding_state_fetch'
+require 'messages/service_binding_create_message'
 
 module VCAP::CloudController
   class ServiceBindingCreate
