+++ app/controllers/services/lifecycle/service_instance_purger.rb
@@ -1,4 +1,4 @@
-require 'actions/service_binding_delete'
+require 'actions/v2/services/service_binding_delete'
 
 module VCAP::CloudController
   class ServiceInstancePurger
